\name{dmr.data-package}
\alias{dmr.data-package}
\alias{dmr.data}
\docType{package}
\title{A package with example tiny datasets from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}
\description{A package with example tiny functions from Chapter 1 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}
\details{See examples in Chapter 1.}

\author{
Pawel Cichosz <p.cichosz@elka.pw.edu.pl>
}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\keyword{package}
\keyword{arith}
\keyword{models}

\seealso{
\code{\link[dmr.util:dmr.util-package]{dmr.util}}
}

\examples{
}
