\name{weather}
\alias{weather}

\docType{data}

\title{An example dataset from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}

\description{An example dataset from Chapter 1 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\usage{See Section 1.3, Example 1.3.1.}

\details{See Section 1.3, Example 1.3.1.}

\source{Dataset introduced by Quinlan, J. R. (1986). Induction of Decision Trees. Machine Learning, 1:81-106.}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\examples{
data(weather)
summary(weather)
}

\keyword{datasets}
