\name{weatherr}
\alias{weatherr}

\docType{data}

\title{An example function from the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R}
\description{An example function from Chapter 1 of the book Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. See Appendix B or http://www.wiley.com/go/data_mining_algorithms for more details.}

\usage{See Section 1.4, Example 1.4.1.}

\details{See Section 1.4, Example 1.4.1.}

\source{A modified version of the dataset introduced by Quinlan, J. R. (1986). Induction of Decision Trees. Machine Learning, 1:81-106.}

\references{Cichosz, P. (2015): Data Mining Algorithms: Explained Using R. Wiley.}

\examples{
data(weatherr)
summary(weatherr)
}

\keyword{datasets}
